ended

+++ app/Http/Controllers/EmployeeViewController.php
@@ -9,6 +9,8 @@ use Illuminate\Contracts\View\View;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB as FacadesDB;
+use Illuminate\Support\Facades\File;
+use PDF;
 
 class EmployeeViewController extends Controller
 {
@@ -249,18 +251,18 @@ class EmployeeViewController extends Controller
 
     public function end()
     {
-            $employees = Employee::whereHas('employee_contracts', function ($query) {
-                $query->whereNotNull('start_date')->whereNotNull('end_date');
-            })
+        $employees = Employee::whereHas('employee_contracts', function ($query) {
+            $query->whereNotNull('start_date')->whereNotNull('end_date');
+        })
             ->with(['employee_contracts', 'user' => function ($query) {
                 $query->select('id', 'name', 'last_name');
             }])
-            ->get(['id', 'user_id']); // Retrieve 'user_id' column
-        
-        $employeeIds = $employees->pluck('user_id')->toArray(); // Use 'user_id' column instead of 'employee_id'
-        
+            ->get(['id', 'user_id']);
+
+        $employeeIds = $employees->pluck('user_id')->toArray();
+
         $users = User::whereIn('id', $employeeIds)->get();
-        
+
         $employeesWithUsers = $employees->map(function ($employee) use ($users) {
             $user = $users->firstWhere('id', $employee->user_id);
             if ($user) {
@@ -268,19 +270,41 @@ class EmployeeViewController extends Controller
                 $employee->name = $user->name;
                 $employee->last_name = $user->last_name;
             }
+
             return $employee;
         });
-        
+
         return view('endcontract', compact('employeesWithUsers'));
     }
-    
-    public function determine($employee)
-    {
-        $employee = Employee::findOrFail($employee);
-        $contract = $employee->employee_contracts();
-        $contract->delete();
 
-        return redirect()->back()->with('success', 'Contract determined.');
+    public function determine($employeeId)
+    {
+        $employee = Employee::findOrFail($employeeId);
+        $contract = $employee->employee_contracts()->first();
+        $user = User::findOrFail($employee->user_id);
+
+        $employee->user = $user;
+        $employee->name = $user->name;
+        $employee->last_name = $user->last_name;
+
+        $pdf = PDF::loadView('cvier', ['employee' => $employee, 'contract' => $contract]);
+
+        $pdfPath = public_path('pdf_files/pdf_file.pdf');
+        $pdfDirectory = dirname($pdfPath);
+
+        if (! File::isDirectory($pdfDirectory)) {
+            File::makeDirectory($pdfDirectory, 0755, true, true);
+        }
+
+        $pdf->save($pdfPath);
+
+        if ($contract) {
+            $contract->delete();
+
+            return response()->download($pdfPath)->deleteFileAfterSend(true);
+        } else {
+            return redirect('/home')->back()->with('error', 'No contract found.');
+        }
     }
 
     public function renew($employee)
